backend/requirements.txt
Django<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=23.5
Pillow>=10.0
python-decouple>=3.8
mysqlclient>=2.2.0
